+++ NEWS.md
@@ -1,4 +1,8 @@
 # hellmer 0.1.0
 
+## New features
 * Initial CRAN submission.
 * Removed `structured_data()` method as `texts()` now handles structured data responses.
+
+## Experimental features
+* Development version now supports LLM-as-a-judge refinement for structured data extractions via the `judgements` parameter.